Limit the initial size of the window (trac #53).
We do this by telling the icon view to resize its children and asking
it how big it now is.  If this exceeds the maximum height we allow, we
set the width to the maximum we will allow and ask it to resize again.

After this, if the new size is too large, we limit it to the maximum
size in both dimensions.

	* system-config-printer.py (GUI.__init__): Limit the initial size
	of the window (trac #53).