Use Monitor.update() where appropriate.
Some places need to wait for the update to appear so that they can
perform further actions, such as selecting a newly created printer.

For actions that don't need to do this, but know that they will
trigger a notification, self.monitor.update() is the correct way to do
this.

	* system-config-printer.py (GUI.on_enabled_activate): Tell the
	monitor to check for updates.
	(GUI.on_shared_activate): Likewise.
	(GUI.save_printer): Likewise.
	(GUI.rename_printer): Likewise.
	(GUI.on_copy_activate): Likewise.
	(GUI.on_delete_activate): Likewise.