Added job-related functions to cupspk.
here is a patch which add job-related functions to system-config-printer
(cups-pk-helper branch).
PolicyKit versions of these functions are in CVS of cups-pk-helper.
It also change the way the Cancel button in a PolicyKit dialog is
handled. Because this state is not distinguishable from policy saying
no or from 3 wrong passwords. The ObtainAuthorization returns FALSE for
the 3 cases. An info dialog pop-ups in such a case.