Revert last change.  Reasons below.
If 'hp-info -d' requires a CUPS queue it seems to me that it is a bug
with hp-info.  We can't guarantee the format, location, or existence
of models.dat as that is a data structure internal to HPLIP.

We *MUST NOT* override the user's choice of device URI under *ANY*
circumstances as this makes it absolutely impossible to debug backend
problems.  Instead, all possible device URIs should be listed in the
>Connection list and the most appropriate one pre-selected.