Reverted commit b72a78e.  Reasons below.
There were some bug-fixes that applied to 1.0.x in this commit.  I
have applied them, separately, to the 1.0.x branch, to be merged into
this branch.

The hp:// backend selection still does not work in the correct way.
It needs to be shown in the list of available connections for the
physical device, i.e. in the 'Connection' list in the bottom
right-hand side of the window.

If work needs to be done to determine whether this URI is available,
it should be done when the cursor is changed to that physical device
for the first time.