udev-configure-printer: fixed removal when cupsd not running.
If the queue for a device has been removed, then cupsd is stopped, and
then the device is disconnected, reconnecting the device once cupsd
has been re-started fails to add a queue.

The fix is always to remove the usb-uris line for the device on
disconnection even if we cannot connect to cupsd.

* udev/udev-configure-printer.c (cups_connection): Removed.
(find_matching_device_uris): Exit if unable to connect to cupsd.
(for_each_matching_queue): Return early if unable to connect.
(enable_queue, disable_queue): Likewise.
(main): Set the CUPS password callback here.