Suggest installing foomatic-db-ppds when appropriate.
* cupshelpers/cupshelpers.py (parseDeviceID): Don't set the CMD
list to a list of the empty string when not present; set it to
the empty list.
* system-config-printer.py (NewPrinterGUI.nextNPTab): After the
device is selected decide whether it might support PostScript.
If it might, and foomatic-db-ppds is not installed, suggest
installing it.

NOTE: The application will have to be restarted for the new
PPDs to be discovered.

This is not particularly elegant.  It will be fixed properly in
1.2 by installing packages by MFG/MDL tags.