Use asyncconn for fetching device lists.
This avoids a thread-safety issue: as getDevices is an operation
that (by default) requires authentication, when PolicyKit is
available we will use it.  However, doing so from within a
TimedOperation will cause it to be done from a separate thread.

As PolicyKit is implemented over the D-Bus system bus, and we
are already using the system bus in Monitor, this is not safe.

By using the new asynchronous calling class, direct IPP
connections are done from a separate thread (as there is no
asynchronous interface in libcups) whereas PolicyKit calls
are performed asynchronously in the same thread.  This means
that D-Bus is only ever used from one thread, and so is safe.