Fix crash copying printer with certain job options (bug #554268).
When a job option such as page-border (which has known supported
values) is present on a queue, and that queue is duplicated, we decide
whether to set that job option based on whether it is the system
default.  However, the system default is not set for that type of
option.

* options.py: Add 'get_default' method to optionwidget interface.
* system-config-printer.py (GUI.save_printer): Use it.