Finally fixed race condition in asyncipp.
The IPP connection thread was using the idle loop for callbacks, but
the error/reply callback functions were stored in class instance
variables.  By the time the idle loop ran the callbacks they may be
for a different caller!

The fix is to pass the error/reply callback function itself to the
idle timeout, as well as the actually error/reply data.