Strip model names before " PCL" and " UFR II" for Canon drivers.
Andrew Kisliakov writes:

==>
I work with two products – a locally developed driver product called
CQue which provides PS, PCL and PXL drivers, and a Canon developed
driver for the UFR II language.

[...]

If the two strings “ pcl” and “ ufr ii” were added to this set of
strings, the drivers are then correctly classified. I’ve tested this
and it seems to work.
<==