Silence error messages of missing PrinterDriversInstaller D-Bus service
On Ubuntu the mentioned D-Bus service is not available, and here are
really ugly console message appears:

system-config-printer-applet: failed to start PrinterDriversInstaller
service: org.freedesktop.DBus.Error.AccessDenied: Connection ":1.57"
is not allowed to own the service "com.redhat.PrinterDriversInstaller"
due to security policies in the configuration file

People could think that there is a security problem with
system-config-printer. See also Ubuntu bug #530218.

The change should perhaps be improved, for example let the ./configure
script set a switch whether the PrinterDriversInstaller D-Bus service
is installed or not and only show the message if the service is
installed.