Improvements for setting up Bluetooth printers
1. bluetooth://... URI are explicitly supported now, so on the right
hand side of the window with the list of detected devices a nicer
screen appears and not the screen for a custom URI.

2. Bluetooth devices are searched together with the network devices
and not together with the local devices. Discovering Bluetooth devices
takes considerably longer than discovering local devices (USB or
parallel). This way the local devices pop up in the list immediately
and not only after the long time the bluetooth CUPS backend needs for
its device discovery task.