Fix recognition of remote CUPS queues when setting up an IPP queue
If a queue to be set up with an "ipp://..." URI points to a remote
CUPS queue which has already a PPD/driver on the server, we do not
need to (and should not) search for a PPD/driver to set up
locally. This change fixes a bug to make system-config-printer really
recognize such queues and suppress driver configuration.