Integration of the dnssd CUPS backend and assosiation of DNS-SD names and IPs
Currently, there are two methods for CUPS to discover local printers
in the network: DNS-SD which identifies printers with DNS-SD names and
SNMP which identifies printers with IP addresses. In addition one can
find two different HPLIP URIs for HP's network printers, also one
DNS-SD-name-based and one IP-based. All these must be correctly
associated to the physical printers, so that each physical printer
will appear only once in the list of discovered printers.

This change solves this problem reliably but it could not be avoided
to call the command line utility avahi-browse for that as there are no
well-documented Python APIs to communicate with Avahi. To avoid any
kind of problems in terms of slowness or blocked UI, the call is done
asynchronously and the result is cached so that only one call is
needed for identifying all discovered network printers.

A replacement of the one command line call by an appropriate API call
is welcome.