Reverted second DNS-SD resolver.
Was this even necessary?  If we really need to resolve special HPLIP
device URIs, the dnssdresolve module should be used to do that
asynchronously and at the right point (i.e. not inside PhysicalDevice
but between getting the device URIs and adding the devices, like is
already the case with dnssd:// URIs).