Enhancements of the PhysicalDevice class for the CUPS "dnssd" backend
This is especially to improve association of DNS-SD- and
SNMP-discovered printers, so that the same printer, discovered with
the two methods appears only once in the list of discovered
devices. There are also improvements for getting mmake/model/serial
number info and for network printers the IP address and (if available)
the host names are also listed. This way admins can easily identify
the discovered printers, even if there is more than one of the same
model in the network.