Handle the situation where cupsd has died more gracefully.
NewPrinterGUI's init() function now returns True/False; if it failed,
update to see if the service is still running.

In other exceptional situations, update to see if the service is still
running.

When we get the cups-connection-error signal from the monitor, do not
try to reconnect through libcups by calling getClasses; instead,
immediately go into the unconnected state.