Make automatic driver download also working when changing the driver
When changing the PPD file of an existing print queue ("Make and
Model" in "Properties" dialog) the cupshelpers.getDevices() function
is called with the scheme of the device URI to find the printer's
device ID. The scheme needs to be supplied as ASCII string and not as
UniCode, otherwise the scheme is not recognized and the appropriate
CUPS backend does not get called. The device ID is needed to suggest
the correct drivers and to trigger automatic driver downloads.