Job viewer's attribute window: Convert job numbers and attribute values to strings
In the code for the job attributes window in the job viewer
(right-click a job and choose "View Attributes") on two points
arguments have to be converted to strings. Once for the notebook page
labels as the job IDs are numbers, and second for the attribute
values, they can be numbers or even data structures (Ubuntu bug #733088).