Raised priority for SpliX driver
Raised the priority of the SpliX driver for Samsung laser printers, so
that it is used as the default driver for these printers, and not
"Foomatic/gdi" any more. The driver is free software and is superior
in terms of reliability and printout quality, so it should not be
categorized as a non-free third-party driver (Ubuntu bug #793741).