udev: don't rely on retriggering printers, enumerate them instead
Retriggering udev devices is primarily a tool to make sure changed udev
rules should be applied. It should not be used in any normal codepaths.

This patch makes it unnecessary to retrigger printer devices after
starting CUPS. Instead we simply enumerate them in a service that is
started at the same time as CUPS.

This patch also includes the systemd unit file that makes sure we are
called at the right time.

We use "RemainAfterExit=yes" in the service file to ensure the
enumerator is called only once during runtime.