Fixed bold print of tab labels on option conflicts
If there are conflicting option settings on the tabs "Printer Options"
and/or "Installable Options" of the printer properties dialog, the tab
names at the left of the dialogs should be shown in bold. This patch
makes this working correctly in all situations, especially after
changing the driver/PPD for a print queue.