Improve handling of remote CUPS queues reported via DNS-SD
A CUPS server can also broadcast its queues vis DNS-SD and not via IPP
(like Mac OS X machines with shared printers do). In this case the
queues are not made available automatically on the client, but the
broadcasted queues can be selected in the add-printer wizard under the
auto-detected network devices. This patch improves the labeling so
that the server is listed with its name or IP (and not with the
description of its first queue) and in the "Connection" menu of the
server the queues are shown with their descriptions (and not all
undistinguishable with "Remote CUPS printer via DNS-SD").