On asynchronous IPP connections make sure that the password dialog is repeated if a wrong password is entered (Ubuntu bug 653132).
On asynchronous IPP connections there is no "Not authorized" pop-up
and no reappearing opf the login/password dialog to try again if one
enters a wrong password. It gets fixed when removing the
"self._try_as_root = False" in before the reconnection attempt as
root. There cannot happen an infinite loop due to this anyway, as the
try-as-root code also checks the user name and does not reconnect as
root again if the user is already root.