Allow printing test page on remote printer without PPD
If a queue pointing to a remote CUPS queue (for example on a server
which does not broadcast) is created, this queue is usually created as
a raw queue so that the driver on the server gets used. Due to the
fact that the queue is raw, system-config-printer does not allow to
print a test page on it and therefore does nothing when clicking on
the test page button in the properties dialog. This patch allows
printing a test page on such queues (no PPD file, remote queue, not a
discovered queue). In addition, the user is also asked whether he
wants to print a test page when he creates such queues, and for queues
on which one really cannot print a test page (local raw queues or
discovered remote raw queues) the test page button is not shown.