Use hpcups instead of ps with HP LaserJet 12xx/13xx
Use the PCL-5e driver (HPLIP's hpcups) instead of PostScript as
recommended driver on the HP LaserJet 12xx and 13xx as they do not
come with enough memory for PostScript. The PostScript driver options
are still listed for manual selection so that users with a memory
extension in the printer can use them (Ubuntu bug #872991).