Added debugprint() calls to more easily debug problems of driver assignment.
The extra debug info shown are the lists of valid driver types (from
preferreddrivers.xml), the assignment of the driver types to the
suitable PPDs and the resulting priority order of the PPDs. Lists are
sorted by the priority order defined in preferreddrivers.xml. This
helps debugging problems with preferreddrivers.xml and improving it.