Added driver auto-download functionality to scp-dbus-service
This patch adds the functionality to auto-download printer drivers
as LSB-based packages via the OpenPrinting web site also to the
process of finding the best printer driver in the D-Bus service
scp-dbus-service. This way the auto-download also works if instead
of the native GUI of system-config-printer another frontend GUI
like the printer setup tool in GNOME Control Center is used.

The patch also adds support for download of signed PPD-only
packages, adds debug output to the install-printerdriver script,
support for introduction of package signing on the server and for
change of keys, and fixes downloading unsigned PPD-only packages.