Fixed a crash and an infinite loop occuring on driver download from OpenPrinting
There are two problems which can happen when a detected printer is not
locally supported and OpenPrinting is searched for it with a positive
result (like Canon Pixma MP160). First, there is a crash in the thread
taking care of inbteractive control of the new printer dialog, making
"Cancel" not work any more when the driver download is
presented. Second, after the download sytem-config-printer can get
stuck in an infinite loop, not stopping when the
"install-printerdriver" helper script terminates.

This change solves both problems.