Fix possible race condition leading to a false error alarm after driver package download
Fix for the earlier committed fix of the infinite loop after
downloading a driver package via OpenPrinting: Now the loop is
terminated when install-printerdriver outputs "done" but perhaps
somewhat early (race condition?) as this causes a false alarm of an
error exit of install-printerdriver, telling to the caller that the
driver was not installed, letting the caller not reload the system's
PPD list and so the freshly installed driver does not get found by the
subsequent print queue setup. Now we ignore an error condition after
install-printerdriver has output "done". With this we should have a
working fix for Ubuntu bug #1248303.