Create filelist of freshly installed driver packages in UTF-8
When installing a printer driver package, the helper script
install-printerdriver returns a list of the freshly installed files to
make it easier to assign a PPD of this package on the following
creation of a print queue. When we read it, we create the file list
array directly as UTF-8 strings for easier usage later on.