Make inter-process communication with install-printerdriver script more reliable
For adding package repositories and installing printer driver packages
out of them the install-printerdriver script is used. It provided data
to the caller from both stderr and stdout. As the data could come out
in an unexpected order a buffer could run full (especially with the
openprinting-gutenprint package which has a very long file list) and
block system-config-printer or make it crash (broken pipe).

Now all relevant data (progress and file list) is provided via stdout
and only stdout is read. This way the install-printerdriver script
works reliably.