Support license text of driver on separate, downloadable URL
The OpenPrinting web server does not only allow the driver's license
texts being embedded in the Foomatic driver XML file but also having
the text on a separate URL and link to it in the driver XML file. This
makes it easier especially for printer manufacturers to maintain their
often long license texts and also translations for them.

system-config-printer supported only directly embedded license texts
and as Epson switched to separately downloadable ones, their drivers
could not get downloaded any more.

This patch adds support for the separate texts.