Make download_gpg_fingerprint() find CA certificates in Fedora and RHEL
Fedora and RHEL store certificates in a different place than Debian and
Ubuntu, so we need to add those paths here so that signed packages
downloaded from OpenPrinting.org can be properly validated.

Also, we want to use the first CA certificate found, so added a break in
the for look when the first file is found.