Replace GtkTable for downloadble driver's details with a GtkGrid
Besides being deprecated widgets, using a GtkTable here does not
work as nicely as a GtkGrid: it makes all its cells to require way
more space than they need for some reason and then filling that
space, causing a very weird visual effect, wasting lot of space and,
even worse, forcing the GtkWindow to be bigger than what it should be.

So, migrate this table to GtkGrid, so that we can keep the functionality
and the information offered by this window, without the weird effects.