Make the downloadable drivers page usable in lower resolutions
Reorganized how widgets are presented to the user, so that only
the strictly needed (License terms + Accept/Reject buttons) are
always shown, with the rest of the additional information hidden
behind expander widgets.

This is but necessary if we want users to be able to install
drivers when running on very low resolutions (e.g. 720x480).