Changed the default size for the NewPrinter window to 800x600
720x480 is the minimum resolution used when in composite mode (NTSC), so we
use a slightly bigger resolution mode than that to make sure that this window
will use at least all available space when running in those low resolutions.

Note that this is not setting a minimum, so the window can be still resized down
if desired. Also, when launching the window in lower resolutions, the window
will simply launch maximized, so it should be all good.