Make the main content of the NewPrinterWindow window scrollable
We leave the buttons of the action area out of the scroller to prevent
confusion, and also set a reasonable minimum (480x340 that should work
fine even in extremely low resolutions.

This is the last step towards making the NewPrinterWindow, and specially
the "downloadable drivers" page usable in lower resolutions, adding a
backup plan for when the window does not fit even after doing all the
adjustments, so that the user can still access to the window's content
and reach the buttons in the action area.