Prioritize IPP in the printer connection type list.
As IPP is the most modern and sophisticated network printing protocol
and especially supports driverless printing, let IPP options show first
in the list of possible connections for a discovered network printer.
Only HPLIP stays more highly prioritized as for HP multi-function
devices as HPLIP print queue is required for the scanner to work.