When adding a printer, consider also the "driverless" and "bjnp" CUPS backends as network backends
When starting the wizard for adding a new printer the available
printers and CUPS backends are enumerated by calling all CUPS backends
in discovery mode, but in contrary to CUPS' "lpinfo -v" the bunch is
devided up into two groups, first the faster local printer backends
are called and the appropriate printers listed and after that the
slower network printer backends. This way the local printers appear
quickly in the wizard and the user can already proceed when the
desired local printer shows up.

The "driverless" and "bjnp" backends were missing in the list of
network backends and therefore were considered local, ending up with
the local backend round taking too long.