Suppress backend-only entry for the "bjnp" backend in the list of discovered printers
The list of discovered printers and CUPS backends in the wizard for
adding a new printer has an entry for the "bjnp" CUPS backend (backend
for proprietary network protocol of some Canon printers). This entry
appears when no appropriate Canon printer gets found and the entry in
the list is useless. Therefore we suppress it.