When adding a device to the list of discovered devices, use the original physical device entry
In the new-printer wizard the list of discovered printers in the first
wizard step is built up in several stages. To avoid duplicates in each
stage the devices are not only added to the already existing physical
device record but also to extra physical device record to mark them as
new entries which still need to get added to the list.

Before, these new physical device records were used for generating the
list entry. Now we look up the original physical device entries as
they can contain more information (especially IP address and host
name) for the list entry.

In addition, som extra debug information is logged.