Consider entries for HP printers as the same physical device if the serial number matches
If an HP multi-function device on USB has also fax functionality,
there are also discovery entries for the "hpfax" backend of HPLIP and
an additional entry for the "usb" CUPS backend. These entries have
model names which differ from the printer's model name, leading to
extra entries in the list of discovered printers. To join these
entries we consider two HP printers as equal if they have the same
serial number, regardless of the model name.