When listing discovered USB devices, distinguish between printer and fax
Some USB printers appear as two USB devices, one for printing and one
for fax. This would lead to 2 connection types in the connection type
list which are names "USB" and undistinguishable. Now we check on USB
printer entries whether they are for fax and mark them as "Fax - USB".