Add checkbox for sharing when adding a printer
In the printer add menu there is no checkbox or any other
way to specify whether to use shared access or
not when adding a device. The device is set to the default
value defined in CUPS after adding.

The patch adds a new line to the "Describe Printer" section,
containing a checkbox that allows users to specify whether to
use shared access or not for the device being added